move qa_qc to its own module update tests and imports update docs
