+++ docs/source/reV/reV.qa_qc.summary.rst
@@ -1,0 +1,7 @@
+reV.qa\_qc.summary module
+=========================
+
+.. automodule:: reV.qa_qc.summary
+   :members:
+   :undoc-members:
+   :show-inheritance:
