+++ docs/source/reV/reV.utilities.cli_dtypes.rst
@@ -1,0 +1,7 @@
+reV.utilities.cli\_dtypes module
+================================
+
+.. automodule:: reV.utilities.cli_dtypes
+   :members:
+   :undoc-members:
+   :show-inheritance:
