+++ docs/source/reV/reV.utilities.rst
@@ -11,6 +11,7 @@ Submodules
 
 .. toctree::
 
+   reV.utilities.cli_dtypes
    reV.utilities.curtailment
    reV.utilities.exceptions
    reV.utilities.slots
