+++ reV/utilities/__init__.py
@@ -2,4 +2,3 @@
 """
 reV utilities.
 """
-from .qa_qc import Summarize, SummaryPlots
